feat(cfapi): include origin-ca-issuer version
The team operating the Origin CA have requested this issuer include
version information in the User-Agent to aid in debugging issues on
their side and understand the rate at which updates to the issuer have
been adopted.

Introduces an internal "version" package where the binary version is
set via LDFLAGS. Starting with Go 1.24 this information will be embedded
into the binary's module, which can be queried instead.